Rename lfire as pylfire
