--- lfire/lfire/__init__.py
@@ -1,3 +1,0 @@
-# flake8: noqa
-
-from lfire.methods.lfire import *
