+++ pylfire/pylfire/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from pylfire.methods.lfire import *
