+++ pylfire/setup.cfg
@@ -1,0 +1,9 @@
+[flake8]
+max-line-length = 100
+
+[tool:pytest]
+filterwarnings = 
+    ignore::DeprecationWarning
+    ignore::FutureWarning
+    ignore::UserWarning
+    ignore::RuntimeWarning
